use long type for stream version (#28)

+++ OpenFTTH.EventSourcing/IEventStore.cs
@@ -7,11 +7,11 @@ namespace OpenFTTH.EventSourcing
 {
     public interface IEventStore
     {
-        void AppendStream(Guid streamId, int expectedVersion, object[] events);
-        Task AppendStreamAsync(Guid streamId, int expectedVersion, object[] events);
+        void AppendStream(Guid streamId, long expectedVersion, object[] events);
+        Task AppendStreamAsync(Guid streamId, long expectedVersion, object[] events);
         void AppendStream(IReadOnlyList<AggregateBase> aggregate);
         Task AppendStreamAsync(IReadOnlyList<AggregateBase> aggregate);
-        object[] FetchStream(Guid streamId, int version = 0);
+        object[] FetchStream(Guid streamId, long version = 0);
         IProjectionRepository Projections { get; }
         IAggregateRepository Aggregates { get; }
         ICommandLog CommandLog { get; }
